EN-2137 Move auth test helpers to a separate file.
